Add config option to link external accounts with existing accts
If QA_LINK_EXTERNAL_USERS is true, then when you first log in with
an external account (e.g. LDAP), if that account has the same email as
an existing account, then the LDAP login is linked to the existing
account.

If QA_LINK_EXTERNAL_USERS is false, the existing behaviour is retained,
which is to attempt to raise the email address as an error. In the case
of LDAP, that fails horribly in a redirect loop